made user's grading cinema func

+++ Myproject4/database/migrations/2020_05_13_013518_create_contribute.php
@@ -19,8 +19,8 @@ class CreateContribute extends Migration
 	    			$table->text('contents');
 	    			$table->string('picture');
 						$table->string('genre');
-						$table->unsignedInteger('satisfaction')->default(0);
-						$table->unsignedInteger('recommended')->default(0);
+						$table->float('satisfaction', 5, 2)->default(0);
+						$table->float('recommended', 5, 2)->default(0);
             $table->timestamps();
         });
     }
